refactor(Alu): split Vset from Alu (#1906)
